Navigation Sibiling Step
Provides navigation for 'stepping' through page siblings one at a time.
This means if there is a following sibling a 'next' link will be display
ed, and if there is a preceding sibling a 'rev' link will be displayed.
This navigation follows HTML5 standards, with correct 'rel', and should
be wrapped in a <nav></nav> element when used.